feat: Teach `valueOf` how to handle maps and collections.
The convenience method `valueOf` would turn primitives into the
corresponding json nodes. Maps and collections however would not be
converted. Which causes the following to fail:

.. code-block::

    Map<String, Object> map = Collections.singletonMap ("key", "value");
    valueOf (map).get ("key").asText ();

The result of ``valueOf (...)`` is a ``PojoNode``, which responds with a
null value to ``get ("key")``. Thus the last call to ``asText ()`` will
cause a nullpointer exceptions.

After this change ``valueOf (...)`` will no longer return ``PojoNode``s,
instead it returns an ``ArrayNode`` for collections and an
``ObjectNode`` for maps.