feat: Add interaction with streams
When we want to collect data represented as a java stream to an
``ArrayNode`` our only option was to first collect it to a list or other
collection and then use ``valueOf``.

This is cumbersome and forces you to allocate an intermediate list
object that is than discarded.

.. code-block:

    Stream<String> stream = Stream.of ("a", "b", "c");
    Json.valueOf (stream.collect (Collectors.toList ());

Instead we can now create a collector that will directly result into an
``ArrayNode``:

.. code-block:

    Stream.of ("a", "b", "c").collect (Json.collectToArray ());

Simularly we have collectors to go from streams to ``ObjectNode``s.