feat: Add convenience stringify and parse methods
It is possible to produce a string from any object (includinng any of
the node types).

.. code-block::

    ObjectMapper mapper = ...;
    Object subject = ...;
    String json = mapper.writeValueAsString (subject);

    // or

    new ObjectMapper ().writeValueAsString (subject);

This method throws an ``IOException`` which is anoying to deal with. The
serialization all happens in memory so real io exceptions should not
happen.

We also need a mapper which would then need to be configured with
modules and time serialization settings. To support easy convertion to
json strings we have add ``strinigfy ()`` and ``prettyStringify ()``
methods to the json class.

Similar issues occur when trying to parse a string in memory. For those
cases we have provided ``parse ()`` which will return a ``JsonNode`` or
any of its subtypes.