feat: Add convenience mapper configuration
``ObjectMapper``s don't work well with more recent java types, there is
a standard library to add the serialization and deserialization. However
these are not present when calling the default constructor. The same
happens with classes from ``java.time``. In addition instants are
serialized as numbers (seconds since epoch) while they are probably
better serialized as iso strings.

Therefore we have added a method that will return a preconfigured
instance: ``Mappers.createBaseMapper ()``.

A second issues is that adding a custom serializer involves a lot of
boilerplate code

.. code-block::

    ObjectMapper mapper = ...;
    SimpleModel module = new SimpleModel ();
    module.addSerializer (Custom.class, new StdSerializer<Custom> (clazz) {
        @Override
        public void serialize (Custom value, JsonGenerator gen, SerializerProvider provider)
        throws IOException {
            /* ... */
        }
    });
    mapper.registerModule (moduel);

While the only pieces of real code there are the class we wish to
serialize and the body of the inner class. This commit adds a functional
interface that captures the essence of the serialization, and some
methods to add configuration to a module and an object mapper

.. code-block::

    ObjectMapper mapper = ...;
    Mappers.configure (mapper, Mappers.module (m -> {
    	Mappers.serialize (m, Custom.class, (value, gen, provider) -> {
	    /* ... */
	});
    });