refactor!: Rework adding JsonException serialization
The ``Mappers`` api assumes that you work with methods that take a
module and configure with additional serializers. The JsonException
class however only has a static inner class that acts as the serializer.

Add a new static method that will play nicely with the ``Mappers`` api.

In the process we have renamed the inner class that implements the
jackson serialization process for json exceptions.

.. rubric:: Breaking Change

The static inner class may have been used directly to configure the
serialization of json exceptions. The name has changed and users should
only need to do the same rename refactoring.

Alternatively they can switch to ``Mappers.createBaseMapper ()``, or use
``JsonException::configure`` in combination with the ``Mappers`` api.